# Expose `reserve()` to `Dictionary`, use it on the GDScript VM