@clayjohn 2025-09-20
> so I thought: why not expose it to `Dictionary`

"Why not?" is never a good reason to add something to Core. While the code is minimal, it has no reason to be exposed to end users in the first place. It just adds another entry to the API that users will never touch. If sizing Dictionaries is a bottleneck, then you are already using the wrong data structure and no amount of calling `reserve()` is going to save you

This is exactly the sort of improvement contemplated by the first two of our contributor best practices:
https://contributing.godotengine.org/en/latest/engine/guidelines/best_practices.html#to-solve-the-problem-it-has-to-exist-in-the-first-place