@DeeJayLSP 2025-09-20
> "Why not?" is never a good reason to add something to Core.

You might have understood things wrong. Dictionary is used in countless places within the engine that can benefit from this.

> While the code is minimal, it has no reason to be exposed to end users in the first place. It just adds another entry to the API that users will never touch.

Except it isn't? It's only exposed for internal use within the engine where beneficial.