@clayjohn 2025-09-20
That's my bad. I didn't notice that there wasn't any binding code touched. That takes away my concern about this being exposed to end users. Adding something like this for internal use is a lot less of a problem.

That being said, if it's truly a benefit. Then this PR should be using it somewhere where dictionary allocations were a problem. Remember, the problem always needs to come first. We don't add solutions to the engine and then plan on finding problems. We start by identifying a problem, then we make changes to solve that problem. 

This may very well be the correct solution to _a_ problem, but in order to merge it, we need to know _what_ that problem is. 

It isn't enough to say that some places _could_ benefit from this change. Show me what places _do_ benefit from this change by profiling them before and after the change.