@Shadows-of-Fire 2025-09-20
> If sizing Dictionaries is a bottleneck, then you are already using the wrong data structure and no amount of calling reserve() is going to save you

That seems like a bit of a nonsensical take on this particular feature. `reserve()` is a standard feature for maps in most languages, and if you know the size AOT you can optimize with a reservation, which can be significant for large dictionaries or for manual copies or transformations of large dictionaries.

I would consider this a core library feature for `Dictionary` that is currently absent. Given that the cost of adding the API surface is effectively zero (as the API is a simple passthrough to the underlying `HashMap`), there would not be sufficient justification to deny it. 

> This may very well be the correct solution to a problem, but in order to merge it, we need to know what that problem is.

You do know what that problem is. Churn in a `HashMap` is wasteful. You don't need to have strict performance numbers for something that is this straightforward.