@clayjohn 2025-09-20
> There are numerous situations within the engine where known final size dictionaries could be built faster for basically free by reserving. Isn't this a problem?

Sure, find one and show me. If there are numerous it should not be hard. I'm really not asking for much. 

> The point is that doing everything inside a single PR would be hard to keep track of.

Sure, don't do everything then. Just solve the low hanging fruit issues. As it stands this PR does nothing, I am asking for it to do _something_ before merging.