@clayjohn 2025-09-20
> That seems like a bit of a nonsensical take on this particular feature. reserve() is a standard feature for maps in most languages, and if you know the size AOT you can optimize with a reservation, which can be significant for large dictionaries or for manual copies or transformations of large dictionaries.

@Shadows-of-Fire Dictionary isn't our standard map in the engine, Hashmap is. By design we don't use Dictionary in any performance sensitive areas. Which is why I said above, if we find Dictionary in a performance sensitive areas, we are likely using the wrong data structure. 

As I said above, if it's so obvious that this will fix a problem somewhere, then just find that problem and point it out. I'm not saying we shouldn't merge this, quite the contrary, I'm saying we _should_ merge it _after_ we have identified the problem it is solving and then used it to solve that problem. 

If the problem is so obvious, you should have no problem finding it. To be blunt, don't waste your time arguing about whether something is useful in theory. I'm certain that 'reserve()' is useful in theory. Instead, spend your time showing where it is useful 'in practice'