@DeeJayLSP 2025-09-20
A few cases I could find in `core/`, performance critical or not:
- `Engine::get_version/author/copyright/donor/license_info()`
- `ProjectSettings::_add_builtin_input_map()`
- `Image::_get_data()`
- `marshalls.cpp` (`decode_variant()`)
- `PListNode::get_value()`
- `ResourceLoaderBinary::parse_variant()`
- `ResourceFormatLoader::rename_dependencies()`
- `Resource::_duplicate_recursive()` and `Resource::_find_sub_resources()`
- `AStarGrid2D::get_point_data_in_region()`
- `TriangleMesh::intersect_*_scriptwrap()`
- A few functions within `Object`
- `Script::_get_script_constant_map()` and `ScriptServer::save_global_classes()`
- `OS::get_memory_info()`
- Most functions in `Time`
- `TranslationPO::_get_messages()`
- `Translation::_get_messages()`

The one case I applied `reserve()` in this PR speeds up Dictionary construction from GDScript. And no, we can't use HashMap there.