@Ivorforce 2025-09-20
@clayjohn Growth is especially costly for hashing types, so `reserve` is especially useful. The PR currently optimizes the case of constructing dictionaries in GDScript, which sees a 8-15% gain from GDScript. This is already practically useful and motivates the PR to be merged in its own right.

In general though, while we don't use `Dictionary` internally, we can expect this function to be useful in situations where we have to construct `Dictionary` for the user (e.g. for function returns).
The main alternative would be offering copy / move constructors, but those are not always feasible.