@Mickeon 2025-09-20
"Expose" is a very strong word in Godot, as you can see. I would suggest renaming the commit and title to something else. Maybe "_Allow `Dictionary` to use `reserve()` in the GDScript VM_"?