@DeeJayLSP 2025-09-20
`HashMap` has a constructor where you can input the initial capacity. Would it be wise to add one to `Dictionary` too?