@AThousandShips 2025-09-20
I'd keep that separately, that would only work for the C++ side due to overrides etc. so it's less critical, we don't have it for `LocalVector` either, it just saves one line